additions
hi, I did everything, but it will be cool if you take a look on it to see if this is wat you were thinking.
the app receive a word and look for a flower in the list that contain that word in one of its attributes.
it returns a list of flowers, and it print it in a list box with all the attributes.
if you want to add somethings or make changes, or if you dont like at all of the way that i realize it, you can change.
thanks